Corfu Incident Event Plus unitstacks.txt from August 8th
Unfortunately no one had an earlier back up of the unitstack.txt so I
had to upload one from August 8th in order to fix whoever screwed up all
the vps. This upload mostly contains things having to do with the Corfu
Incident which may or may not fire properly on the date. I have not
added localisation for the several events the Corfu Incident ecompuses
yet.