better naming scheme.
Replaced the imperial conference with a dynamic event chain. Will take
any province with a given flag, as well as puppets and invite them to
vote. Those that vote yes will be puppeted in the first option, or
annexed in the second. Demonstrates the ability to grab a dynamic number
of provinces, and affect policy. Attaching to an on war declaration,
would let this effect do things like embrago and economic sanctions, or
even form a dynamic alliance to respond to a war.